test: separate claims tests
