+++ test/claimsTests/claims.testSuite.ts
@@ -1,0 +1,5 @@
+import { selfsignedClaimsTests } from "./selfsignedClaimsTests";
+
+export const claimsTests = () => {
+  describe("Selfsigned claim tests", selfsignedClaimsTests);
+};
